src/email_generation.py
```python
import openai
from src.utils import EmailSchema

# Initialize the GPT-3.5 model
openai.api_key = 'your-api-key'

def generate_email(donor, template):
    """
    Function to generate a personalized email for a donor using GPT-3.5
    """
    # Prepare the prompt for GPT-3.5
    prompt = f"{template}\n\nDonor Name: {donor['name']}\nPrevious Donations: {donor['donations']}\nAlignment with Charity's Mission: {donor['alignment']}"

    # Generate the email content using GPT-3.5
    response = openai.Completion.create(engine="text-davinci-003", prompt=prompt, max_tokens=150)

    # Prepare the email
    email = EmailSchema(
        recipient=donor['email'],
        subject=f"Support Our Cause - {donor['name']}",
        body=response.choices[0].text.strip()
    )

    return email
```